Be/272/미션 통계 추가 2/change

See merge request s09-fintech-finance-sub2/S09P22A409!165

+++ back/seesaw/src/main/java/com/doacha/seesaw/model/dto/mission/DailyTopSpendingResponse.java
@@ -12,6 +12,6 @@ import lombok.NoArgsConstructor;
 public class DailyTopSpendingResponse {
     private String dailyTopSpender;
     private Long dailyTopSpending;
-    private int dailyTopSpendingNum;
+    private Long dailyTopSpendingNum;
 
 }
